Added reshadow
