+++ .babelrc
@@ -1,3 +1,8 @@
 {
-  "presets": ["next/babel"]
+  "presets": ["next/babel"],
+  "plugins": [
+    ["reshadow/babel", {
+      "files": "\\.css$"
+    }]
+  ]
 }
