+++ components/Button/index.js
@@ -1,0 +1,8 @@
+import styled from 'reshadow';
+import styles from './styles.css';
+
+export default ({type, onClick, children}) => styled(styles)(
+    <button type={type} onClick={onClick}>
+        {children}
+    </button>
+);
