+++ components/Snippet/index.js
@@ -1,0 +1,8 @@
+import styled from 'reshadow';
+import styles from './styles.css';
+
+export default ({children}) => styled(styles)(
+    <div>
+        {children}
+    </div>
+);
