+++ package.json
@@ -6,9 +6,11 @@
     "test": "jest"
   },
   "dependencies": {
+    "classnames": "^2.2.6",
     "next": "latest",
     "react": "latest",
     "react-dom": "latest",
+    "reshadow": "0.0.1-alpha.65",
     "url-parse": "^1.4.7"
   },
   "devDependencies": {
