+++ pages/index/global.css
@@ -1,0 +1,32 @@
+* {
+    box-sizing: border-box;
+    font-size: 14px;
+    font-family: Helvetica, Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    color: #111111;
+}
+
+::selection {
+    background-color: #79FFE1;
+}
+
+:active,
+:focus:hover {
+    outline: none;
+}
+
+:focus {
+    outline-color: #0366D6;
+    outline-width: 2px;
+    outline-style: solid;
+    outline-offset: 1px;
+}
+
+:global(html),
+:global(body) {
+    width: 100vw;
+    overflow-x: hidden;
+    padding: 0;
+    margin: 0;
+}
