+++ pages/index/index.js
@@ -1,7 +1,8 @@
-import { useState, useEffect, Fragment } from 'react';
+import { useState, useEffect } from 'react';
 import Head from 'next/head';
 import { useRouter, Router } from 'next/router';
 import UrlParser from 'url-parse';
+import styled from 'reshadow';
 
 import TextField from '../../components/TextField';
 import Snippet from '../../components/Snippet';
@@ -10,43 +11,8 @@ import Table from '../../components/Table';
 import { compareQueryParamsInUrls } from '../../helpers/url';
 import Footer from '../../components/Footer';
 
-function GlobalStyles() {
-    return (
-        <style jsx global>{`
-            * {
-                box-sizing: border-box;
-                font-size: 14px;
-                font-family: Helvetica, Arial, sans-serif;
-                margin: 0;
-                padding: 0;
-                color: #111111;
-            }
-            
-            ::selection {
-                background-color: #79FFE1;
-            }
-            
-            :active,
-            :focus:hover {
-                outline: none;
-            }
-            
-            :focus {
-                outline-color: #0366D6;
-                outline-width: 2px;
-                outline-style: solid;
-                outline-offset: 1px;
-            }
-
-            html, body {
-                width: 100vw;
-                overflow-x: hidden;
-                padding: 0;
-                margin: 0;
-            }    
-        `}</style>
-    );
-}
+import styles from './styles.css';
+import global from './global.css';
 
 const FIRST_URL_PLACEHOLDER = "first.com?alpha=1&beta=2";
 const SECOND_URL_PLACEHOLDER = "second.com?beta=3&gamma=2";
@@ -156,8 +122,8 @@ const IndexPage = () => {
         });
     };
 
-    return (
-        <Fragment>
+    return styled(styles)(
+        <>
             <main>
                 <Head>
                     <title>Compare Urls</title>
@@ -191,7 +157,7 @@ const IndexPage = () => {
                         rows={1}
                     />
 
-                    <div className="buttonGroup">
+                    <div className={styles.buttonGroup}>
                         <Button type="submit" onClick={onSubmit}>Compare</Button>
                         <Button type="reset" onClick={onReset}>Clean</Button>
                     </div>
@@ -211,47 +177,9 @@ const IndexPage = () => {
                         <Table titles={['Param', 'Value']} values={equal} />
                     </Snippet>
                 )}
-
-                <GlobalStyles />
-                <style jsx>{`
-                    main {
-                        display: flex;
-                        flex-direction: column;
-                    
-                        padding: 24px;
-                        max-width: 100%;
-                    }
-                    
-                    h1, h2 {
-                        font-weight: bold;
-                        font-size: 18px;
-                        margin-bottom: 12px;
-                    }
-
-                    .buttonGroup {
-                        display: flex;
-                        margin-top: 24px;
-                    }
-                    
-                    .info {
-                        color: #666;
-                    }
-                    
-                    @media screen and (max-width: 599px) {
-                        main {
-                            padding: 12px 8px;
-                            width: 100%;
-                            overflow-x: hidden;
-                        }
-                        
-                        .buttonGroup {
-                            margin-top: 16px;
-                        }
-                    }
-                `}</style>
             </main>
             <Footer />
-        </Fragment>
+        </>
     );
 };
