+++ pages/index/styles.css
@@ -1,0 +1,34 @@
+main {
+    display: flex;
+    flex-direction: column;
+
+    padding: 24px;
+    max-width: 100%;
+}
+
+h1, h2 {
+    font-weight: bold;
+    font-size: 18px;
+    margin-bottom: 12px;
+}
+
+.buttonGroup {
+    display: flex;
+    margin-top: 24px;
+}
+
+.info {
+    color: #666;
+}
+
+@media screen and (max-width: 599px) {
+    main {
+        padding: 12px 8px;
+        width: 100%;
+        overflow-x: hidden;
+    }
+
+    .buttonGroup {
+        margin-top: 16px;
+    }
+}
